ABC-fw: Adding sample west.yml file.
This "west.yml" file pulls all the necessary components from upstream and creates a
usable zephyr repo for the user.

For now, this has been directly copy-pasted from the zephyr repo, but we will eventually
make changes so that it points to the correct zephyr repo (and the correct branch) among
other things.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>